fix(ui): 修正推理时长位置与完成后不显示的问题
- 修复 sameMessage 未比较 elapsed_ms/turn_status：导致轮次结束后用户消息
  携带的新时长/状态被判定为「未变化」而丢弃，回复末尾复制按钮旁不显示总时长
- 撤销 MessageInput 状态栏（workspace 右侧）误加的推理时长
- 推理中实时时长移到「执行状态指示器左侧」（MessageList）：
  · 流式回复阶段：时长显示在流式回复上方
  · 规划/执行阶段（无流式）：时长显示在 Loader + 状态文案左侧
  最终总时长仍由回复末尾（复制按钮旁）持久展示
